Fix the DriverRepo

+++ src/main/java/com/example/brewquest/controllers/DriverController.java
@@ -4,19 +4,22 @@ package com.example.brewquest.controllers;
 
 import com.example.brewquest.models.Driver;
 import com.example.brewquest.models.User;
-import com.example.brewquest.repositories.Driver_repository;
+import com.example.brewquest.repositories.DriverRepository;
 import com.example.brewquest.repositories.UserRepository;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 
 @Controller
 public class DriverController {
-    private final Driver_repository driversDao;
+    private final DriverRepository driversDao;
     private final UserRepository usersDao;
 
-    public DriverController(Driver_repository driversDao, UserRepository usersDao) {
+    public DriverController(DriverRepository driversDao, UserRepository usersDao) {
         this.driversDao = driversDao;
         this.usersDao = usersDao;
     }
@@ -45,7 +48,7 @@ public class DriverController {
     public String showEditForm(@PathVariable long id, Model model) {
         if (driversDao.findById(id).isPresent()) {
             Driver driverToEdit = driversDao.findById(id).get();
-            model.addAttribute("driver", driverToEdit);  // Make sure the attribute name is "driver"
+            model.addAttribute("driver", driverToEdit);
         }
         return "Driver/edit-driver";
     }
@@ -64,7 +67,10 @@ public class DriverController {
 
     @PostMapping("/driver/{id}/delete")
     public String deleteDriver(@PathVariable("id") long id) {
-        driversDao.deleteById(id);
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (user.getId() == usersDao.findById(id).get().getId()){
+            driversDao.deleteById(id);
+        }
         return "users/profile";
     }
 }
